+++ tests/planner/ompl/OMPLTestHelpers.hpp
@@ -237,7 +237,7 @@ public:
     return std::unique_ptr<TestableOutcome>(new DefaultTestableOutcome);
   }
 
-  // Documentatoin inherited
+  // Documentation inherited
   std::shared_ptr<aikido::statespace::StateSpace> getStateSpace() const override
   {
     return mStateSpace;
@@ -319,7 +319,7 @@ public:
     // Create robot
     robot = createTranslationalRobot();
 
-    stateSpace = std::make_shared<StateSpace>(robot);
+    stateSpace = std::make_shared<StateSpace>(robot.get());
     interpolator = std::make_shared<aikido::statespace::GeodesicInterpolator>(
         stateSpace);
 
@@ -363,7 +363,7 @@ public:
     // Create robot
     robot = createTranslationalRobot();
 
-    stateSpace = std::make_shared<StateSpace>(robot);
+    stateSpace = std::make_shared<StateSpace>(robot.get());
     interpolator = std::make_shared<aikido::statespace::GeodesicInterpolator>(
         stateSpace);